PUBDEV-8383: move code to utils
