+++ h2o-algos/src/main/java/hex/rulefit/RuleFit.java
@@ -24,6 +24,7 @@ import java.util.stream.Collectors;
 
 import static hex.genmodel.utils.ArrayUtils.difference;
 import static hex.genmodel.utils.ArrayUtils.signum;
+import static hex.rulefit.RuleFitUtils.consolidateRules;
 
 
 /**
@@ -570,85 +571,6 @@ public class RuleFit extends ModelBuilder<RuleFitModel, RuleFitModel.RuleFitPara
     
     @Override
     public boolean haveMojo() { return true; }
-
-    static Rule[] consolidateRules(Rule[] rules) {
-        //not sure whether I want to overide existing rules or just create new array jut to write to output table
-        //maybe it doesnt matter will solve later when logic will be done
-        for (int i=0; i < rules.length; i++) {
-            rules[i] = consolidateRule(rules[i]);
-        }
-        return rules;
-    }
-    
-    static Rule consolidateRule(Rule rule) {
-        List<Condition> consolidatedConditions = new ArrayList<>();
-        
-        Condition[] conditions = rule.conditions;
-        List<String> varNames = new ArrayList<>();
-        for (int i = 0; i < conditions.length; i++) {
-            if (!varNames.contains(conditions[i].featureName)) {
-                varNames.add(conditions[i].featureName);
-            }
-        }
-        for (int i = 0; i < varNames.size(); i++) {
-            consolidatedConditions.addAll(consolidateConditionsByVar(conditions, varNames.get(i)));
-        }
-        
-        rule.conditions = consolidatedConditions.toArray(new Condition[0]);
-        rule.languageRule = rule.generateLanguageRule();
-        return rule;
-    }
-
-    static List<Condition>  consolidateConditionsByVar(Condition[] conditions, String varname) {
-        List<Condition> currVarConditions = new ArrayList<>();
-        for (int i = 0; i < conditions.length; i++) {
-            if (varname.equals(conditions[i].featureName))
-                currVarConditions.add(conditions[i]);
-        }
-        if (currVarConditions.size() == 1) {
-            return currVarConditions;
-        } else {
-            Condition potentialLessThan = null;
-            Condition potentialGreaterThanOrEqual = null;
-            Condition potentialIn = null;
-
-
-            for (int i = 0; i < currVarConditions.size(); i++) {
-                Condition currCondition = currVarConditions.get(i);
-                if (Condition.Operator.LessThan.equals(currCondition.operator)) {
-                    if (potentialLessThan == null) {
-                        potentialLessThan = currCondition;
-                    } else {
-                        potentialLessThan = potentialLessThan.expandBy(currCondition);
-                    }
-                } else if (Condition.Operator.GreaterThanOrEqual.equals(currCondition.operator)) {
-                    if (potentialGreaterThanOrEqual == null) {
-                        potentialGreaterThanOrEqual = currCondition;
-                    } else {
-                        potentialGreaterThanOrEqual = potentialGreaterThanOrEqual.expandBy(currCondition);
-                    }
-                } else {
-                    assert Condition.Operator.In.equals(currCondition.operator);
-                    if (potentialIn == null) {
-                        potentialIn = currCondition;
-                    } else {
-                        potentialIn = potentialIn.expandBy(currCondition);
-                    }
-                }
-            }
-
-            List<Condition> currVarConsolidatedConditions = new ArrayList<>();
-
-            if (potentialLessThan != null)
-                currVarConsolidatedConditions.add(potentialLessThan);
-            if (potentialGreaterThanOrEqual != null)
-                currVarConsolidatedConditions.add(potentialGreaterThanOrEqual);
-            if (potentialIn != null)
-                currVarConsolidatedConditions.add(potentialIn);
-            
-            return currVarConsolidatedConditions;
-        }
-    }
     
 }
