+++ h2o-algos/src/test/java/hex/rulefit/RuleFitUtilsTest.java
@@ -179,7 +179,7 @@ public class RuleFitUtilsTest extends TestUtil {
             Condition[] conditions = new Condition[] {condition1, condition2, condition3};
 
             Rule rule = new Rule(conditions, 0.032236840575933456, "somevarname");
-            Rule consolidatedRule = RuleFit.consolidateRule(rule);
+            Rule consolidatedRule = RuleFitUtils.consolidateRule(rule);
             assertEquals("(PSA < 14.730077743530273 or PSA is NA) & (DPROS >= 2.5)", consolidatedRule.languageRule);
             
             condition1 = new Condition(6, Condition.Type.Categorical, Condition.Operator.In, -1, new String[] {"ABC", "AAA"}, new int[] {2, 6},"PSA", true);
@@ -188,7 +188,7 @@ public class RuleFitUtilsTest extends TestUtil {
             conditions = new Condition[] {condition1, condition2, condition3};
 
             rule = new Rule(conditions, 0.032236840575933456, "somevarname");
-            consolidatedRule = RuleFit.consolidateRule(rule);
+            consolidatedRule = RuleFitUtils.consolidateRule(rule);
             assertEquals("(PSA in {ABC, AAA, CCC, BBB} or PSA is NA) & (DPROS >= 2.5)", consolidatedRule.languageRule);
             
         } finally {
